Added more tests for the database endpoints of Users and Channels.More tests are still required for the interactions between Users and Channels.

+++ server/lib/services/channelsServices.js
@@ -11,53 +11,59 @@ var _channelModel = _interopRequireDefault(require("../models/channelModel.js"))
 
 var _userServices = _interopRequireDefault(require("./userServices.js"));
 
-var _config = require("../config.js");
+var _config = _interopRequireDefault(require("../config.js"));
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
+let log = _config.default.log;
 /* Add channel from user.
 // Explanation: This creates a new channel and
  adds it to the ownership of the user.
 */
-async function createChannel(userName, userId, channelName) {
-  let channel = null;
 
+async function createChannel(userName, userId, channelName) {
   try {
     let channel = new _channelModel.default({
-      name: channelName
+      name: channelName,
+      members: 1
     });
-    channel = await channel.save(); // Update channel to the user collection.
+    let newChannel = await channel.save(); // Update channel to the user collection.
 
-    channel = await _userServices.default.addChannel(channel.name, channel._id, userId);
-    return channel;
+    await _userServices.default.addChannel(channel.name, channel._id, userId);
+    return newChannel;
   } catch (e) {
-    _config.log.error(e);
+    log.error(e);
+    return e;
   }
 } // Get channel from channel objectId.
 
 
 async function getChannel(channelId) {
-  let channel = null;
-
   try {
-    channel = await _channelModel.default.find({
+    let channel = await _channelModel.default.findOne({
       _id: channelId
     });
+    return channel;
   } catch (e) {
-    _config.log.error(e);
+    log.error(e);
+    return e;
   }
-
-  return channel;
 } // Get channel from channel objectId.
+// It recieves a channelId and a userId for the one deleting it.
 
 
-async function removeChannel(channelId) {
+async function removeChannel(channelId, userId) {
   try {
-    await _channelModel.default.deleteOne({
+    let response = null;
+    await _userServices.default.removeChannel(userId);
+    let channel = await getChannel(channelId);
+    if (channel.members - 1 === 0) response = await _channelModel.default.deleteOne({
       _id: channelId
     });
+    return response;
   } catch (e) {
-    _config.log.error(e);
+    log.error(e);
+    return e;
   }
 }
